Binalyzer/requirements.txt
pyelftools
capstone
lief
angr
